Support SAM interfaces in LE
Adds support to LiveEdit for treating SAM (single abstract method)
interfaces as support classes. Since these interfaces/their
implementations are not specifically tagged in the bytecode, we use a
heuristic.

As of this CL, LE will now treat a class as a support class if all of
the following are true:
- It is an inner class
- It implements a single interface
- It implements a single public non-constructor method

The existing logic for treating lambdas and Compose classes as support
classes is unchanged.

Fixes: 216334710
Test: BasicCompileTest.kt
Change-Id: Ib2d2ede4d72ec624f939b27b62971c94ce8ad17a

GitOrigin-RevId: 05b89743bd843f01e4c9a665ea68554da41c8507