Suggest resource name for Extract Resource String
When extracting a string resource from source code, the code sometimes
suggests a resource name based off of the extracted value and sometimes
doesn't. With this change, there will always be a suggestion for the
name any time there's a value present. (There still won't be a suggested
name for entirely new values being added, where there's no value to
derive the name from.)

The functional changes in this cl are primarily in
KotlinAndroidAddStringResource.kt and
AndroidAddStringResourceAction.java, where a resource name is now
generated. The logic to generate the name from a given value (and
associated tests) is being moved from
AndroidAddStringResourceQuickFix.java to IdeResourcesUtil.kt, so that
it can be accessed from more locations. Note that the key removed from
CreateXmlResourceParameters.kt was only used to set resource names in
tests; now that the resource name is prefilled, that's no longer
necessary.

Test: Included
Fixes: 132749646
Change-Id: If8e9361f0fd210bfd679381f92d3d0a20cbc858c

GitOrigin-RevId: 33090381ae2a36f1d46297c632019cbdd9722253