Fix focus issue in image/icon dialogs
When a ModelWizardStep doesn't override getPreferredFocusComponent, it's
left up to Swing to figure out how to set the focus in between wizard
steps. This can frequently lead to the focus ending up on a help link or
the "cancel" button, both of which tend to be unexpected for users.

This particular case addresses making new image or vector assets.

Fixes: 68313553
Test: added one
Change-Id: I853663768846c00770181a2e60d0963f40bb9ecf

GitOrigin-RevId: cef2c1611bec2042d587f65a3a621a56af567fda