Fix unwanted refresh and inconsistent Preview behaviour
The DesignSurface of a given ComposePreviewRepresentation was not
being activated on the first activation of the representation. This
is not good, as it makes the context to be different between the
first activation and any other re-activation of a
ComposePreviewRepresentation, what could lead to inconsistent
behaviour between different activations.

For example, there is a bug in CodeOutOfDateTracker, that was
causing some unexpected and unnecessary refreshes to happen after
some builds, but only from the 2nd activation of a representation.

This change addresses both problems, the inconsistent behaviour
between activations and the unexpected refrehes mentioned.

Bug: 246963901
Test: added
Change-Id: I8f738488f1191ed04b785f13ca4e454f0455071b

GitOrigin-RevId: 89a812059ad958dea99f35ea65868fb892318ae2