Treat META-INF/*.version as non binary
By default the APK analyzer treats all unknown files
as binary types. This change adds support for having
non binary files and uses that only for META-INF/*.version
files.

Existing tests cannot cover this as the FileType system
is not initialized and requires a rewirite of how this
is tested. Will require a separate change.

Test: Explained above
Fixes: 110233404
Change-Id: I93c1aa97de04cf18b0550ce31ff3f7bf75782fe9

GitOrigin-RevId: 4fa6b204fbbb20ee324d6c61313aac81e8373a5e