Fix flaky test in DeviceViewPanelTest
The main problem was that we did not wait for the
UPDATE_SCREENSHOT_TYPE_COMMAND to finish. And sometimes the command
finished after the animation of going into 3D mode.

Also the timing of the animation was based on real time.
Replaced that with a time that can be controlled in tests.
(Updated Toggle3dActionTest to use the test time as well.)

Fixes: 246371105
Test: DeviceViewPanelTest
Change-Id: I9dcf66f20dce497148894c83a46b50c147c7b46c

GitOrigin-RevId: 115b1267ff2617fe7f85ff4b04442598fcaabd0e