Increase timeout for preview refresh in tests
Some tests in ComposePreviewRepresentationGradleTest are a bit
flaky, and after looking at the logs of some failures, it seems
that usually when the timeout exception happens, a refresh was
being executed, but for some reason it was slower than expected,
and the timeout exception is actually causing an internally caugh
CoroutineCancelled exception for the refresh job.

Also, by executing the tests with some extra logs, it showed that
succesful refreshes usually take around 16 (+-3) seconds, but some
other times it takes more (around 25 seconds). So a timeout of 30
seconds could be too strict.

Bug: 232880163
Test: N/A
Change-Id: I5fe94af28cc7fa0ea2d4d5645a22c4ec8184928c

GitOrigin-RevId: e4c7296a5d302c60f2c9330b3d75d7945c04773c