Reset the selected process each time the user selects a new device
(second attempt)

Disconnecting from the selected process is what the user expects.

In most cases this happens automatically, because on new device
selection, foreground process detection returns the new foreground
process to which we immediately connect.

In some cases, for example if the device is locked when the user
selects it, foreground process detection doesn't return a process
(a locked device has no foreground process), and therefore we need
to manually disconnect from the current process.

Test: DeviceModelTest
Bug: 247875055
Change-Id: I7e8910454122b3738b6b7a1d91b6b58874f2f51c

GitOrigin-RevId: d3b6212bf523503781e84f5a406f2dad91646774