Reset the selected process each time the user selects a new device
Disconnecting from the selected process is what the user expects.

In most cases this happens automatically, because on new device
selection, foreground process detection returns the new foreground
process to which we immediately connect.

In some cases, for example if the device is locked when the user
selects it, foreground process detection doesn't return a process
(a locked device has no foreground process), and therefore we need
to manually disconnect from the current process.

Test: DeviceModelTest
Bug: 247875055
Change-Id: Iaaad3be7cb46e6202aeb4bb7fc1afa6ebbf4cb92

GitOrigin-RevId: 387e8d8a12bd6f095f9c63124fe651c79e8cce66