Handle "Server failed to start" error from transport
When there are two versions of Studio running at the same time,
only the first one can start the Transport server.

When the server can't be started, running the comand that starts
the transport returns the following string: "Server failed to start.
A port number wasn't bound.". (See daemon.cc line 231)

Because this case is not handled in TransportDeviceManager, we keep
trying to start the server forever.

This change detects this error and throws an exception, which results
in publishing `onStartTransportDaemonFail` on the event bus.

Clients of the transport can read the event bus to handle this
error state.

Test: NA
Bug: 244658321
Change-Id: Ie376e974cec5885286f26a9fa1f73c56f6b1e029

GitOrigin-RevId: 6914a55a57a39b64876cddd5ecfe02ceeee845d4