Inline Room stub test setup
Room tests use stub annotations rather than the real thing. A recent
change to the room package has its source converted from Java to Kotlin,
so to accurately test injections we need to be able to have the stubs
implemented in Kotlin as well. But we also still need the existing Java
stubs, since we need to support older versions of the package too.

This change inlines test setup, paving the way to allow parameterization
of whether the stubs are implemented in Java or Kotlin.

Bug: 234612964
Test: Test-only change
Change-Id: I221a54df5a91930974a3a7ce4b620e47bc1134e4

GitOrigin-RevId: 145f8fbff7052165da0f6cdf88f6b6d7807e8fdd