Run RoomSql injection tests with Kotlin source
Room tests use stub annotations rather than the real thing. A recent
change to the room package has its source converted from Java to Kotlin,
so to accurately test injections we need to be able to have the stubs
implemented in Kotlin as well. But we also still need the existing Java
stubs, since we need to support older versions of the package too.

This change adds a stub implementation of the classes in Kotlin, and
adds a variant of the injection tests using that base.

Note that one of the test cases is currently failing due to b/234612964.
I'll submit the fix for that change along with this one, so the test
will be passing when it's submitted.

Bug: 234612964
Test: Test-only change
Change-Id: I7186e6db2c737e58d3f6829a4ce7cd96067995cc

GitOrigin-RevId: 9daf6b7779801821ff3cd094af9836d1ef256329