Fix padding on 'Add String Resource' dialog
This dialog was recently changed with ag/19934008 to use a multi-line
JTextArea instead of a JTextField when adding a string resource, but the
padding on the JTextArea differed slightly from the JTextField. This
caused the resource value text not to line up well with the resource
name. This change rectifies that.

Change-Id: Ib511d65c7157b35f5a1e79a56013e3ab1a1a735c
Bug: 196248641
Test: CreateXmlResourcePanelImplTest.java

GitOrigin-RevId: 83be996b6b36cdb625625d62311d0aedfddee52a