Update benchmarker for variable bitrate.
This change makes it possible for the benchmarker to select and
set the bits per channel and number of latency bits used in the
app dynamically so that we don't need to rebuild the app when
we need these values to change. This includes actually sending
that configuration to the app as well as changing how colors are
extracted and read into integer values from each frame.

Bug: N/A
Test: Included
Change-Id: I04e3399646e1897da41b335924ac05df675f6386

GitOrigin-RevId: 5491e6b8fd8d308083da6172bc9f4d199763b583