[BA] stop passing failed state using fake uiData
This is a small step forward in simplifying this class.
Get rid of fake `failedBuildReportData` that was used only to pass
signal about failure to view creation function. Instead pass lambda
that would create the right view from the places that know what to
create.
Also remove that field from BuildAttributionReportUiData.

Bug: 235469262
Test: existing tests on BuildAttributionUiManager
Change-Id: I62389ea76dc23b88b94b0991b71cd64a1986cd71

GitOrigin-RevId: 46af2722a0299c949337199b72f57334c3f07ec9