Start using AgpVersion where appropriate
Convert many uses of GradleVersion to use the new AgpVersion class
where the surrounding context is clearly working with AGP (e.g. the
version of the plugin applied to a project in Sync; the current
version used to build in the Build Analyzer; the current declared
version in the Upgrade Assistant...)

Not all places have been updated (e.g. in the AGP Upgrade Assistant we
still have plenty of bare GradleVersion types).  Follow-up changes
will be made to specific areas.  A separate follow-up change will
separate GradleVersion from AgpVersion (as it is almost certainly an
error to compare instances of the two classes) before addressing
correctness issues in parsing and comparison.

Bug: 242691473
Test: existing tests updated
Change-Id: I24ab5bd4b8252fa9ebbfbdd6f17ed7b6e6009314

GitOrigin-RevId: d13dacf793416ec40fe20562eb88ccc6e7ae8915