Class static fields obtaining reworked
We used to obtain static fields by calling getDeclaredFields and filtering out all the instance fields, but this approach turned out to have side effects: all the instance field Classes will also be loaded as a result of getDeclaredFields call. It affects the class loading order and may lead to crashes.

In order to fix it now we obtain ids of the fields using JVM TI api, filter out all the non-static fields and get the field values using the GetStaticObjectField jni method.

Test: HeapAnalyzerTest#testGetStaticFieldsNoSideEffect
Bug: 251166138
Change-Id: Ib41cd88820f53ebbbe5d85d72906dde1a9333912

GitOrigin-RevId: 2a53619fdb8ace22f50fc5539969c339699e047e