Add detection of library via app inspection
When the version of a library artifact is not found, also check if
the library is there at all. This will allow app inspection clients
to give a more exact error.

Achieve this by adding expected classNames to LaunchParameters.
When the agent is unable to find the version of a library artifact,
it will also check if the library is loaded by checking if one of
the specified classes exist.

Bug: 246355403
Test: Existing
Change-Id: If5dc815ad266b69971cc3504b26504727c571760

GitOrigin-RevId: 6a74bb541f834eaa029288b9897f5769a77df1aa