Flag "\\}" as redundant within character class
A previous fix removed this flagging, since the escape is not redundant
(ie, it is required) outside of a class. Now that a platform update has
been made, we can go back to flagging this escape as redundant in the
specific case where it occurs within a character class, as in the
expression `"[abc\\}]+"`.

Fixes: 111445868
Test: AndroidRegExpHostTest
Change-Id: Ie089b2f708b954f0a77698f3948e16af1d41d5da

GitOrigin-RevId: 12f82653de1eeee93b1154bf23ae8dad42e8e5a2