Fix protobuf recursion limit for snapshots
For snapshots we call Message.writeDelimitedTo because we need to
write more information to the file. The Message.readDelimitedFrom
needs to read from the input stream before we can construct a
CodedInputStream on which we can specify the recursion limit.

That leads to some code overhead when reading. I had to copy some
of the code from the protobuf library to implement this.
This is not ideal.

Created b/251824432 to eventually find a better solutiion.

Fixes: 251837960
Test: Added unit test
Change-Id: I1ec08234c45cefef7cb1305f64db7d9ed63b51b9

GitOrigin-RevId: bdfc7902b5ca2377774ce4915cd547bf6f686a50