Adapt to generalization of CpuTraceConfiguration
In this change, the studio-side code is adapted to
the proto changes made on base side regarding the
cpu tracing configuration protod. Spefically, the
"CpuTraceConfiguration" message and its children protos.

The base-side changes were simply extractions of the
cpu specific trace configuration protos to a more
general "trace.proto" file, and a removal of the "Cpu"
prefix from these protos. This opens up the protos for
later refactors to further unify the tracing-based protos.

Bug: none
Test: updated
Change-Id: I0db744dafa257801ad089191394ae32595bda304

GitOrigin-RevId: 910f5fcd60b3f1aee10a0441f80f7295f6701080