Create AndroidRunConfigurations with LAUNCH_DEFAULT_ACTIVITY by default
This fixes cases when a source manifest doesn't declare an activity, so hasDefaultLauncherActivity returns false, and we set launch options to DO_NOTHING.

When a user runs configuration we do nothing, it's confusing because we do it without any notification.

This fix restores behavior prior 5b5d5e5 for not-wear projects.

Fixes: 249095314
Test: the project with Activity in library creates configuration with "Default Activity launch" option
Change-Id: I299d7465b82621960a0372c0271b6c211574fdb0

GitOrigin-RevId: ad55da41d4b804b64b467ce6666d63d7dcedf6b4