Remove composable highlighting race condition
Highlighting of composable functions is currently done with an
Annotator, in accordance with JetBrains guidance at
https://plugins.jetbrains.com/docs/intellij/syntax-highlighting-and-error-highlighting.html#annotator.
This runs as a third pass of syntax highlighting, after the parser
phase.

As it turns out, Kotlin actually does its highlighting with a different
extension mechanism. This seems to be running in parallel with the
Annotator, leading to race conditions and non-deterministic behavior for
which highlight "wins".

This change converts the current Annotator into a HighlightVisitor,
which ensures that the built-in Kotlin function highlighting will never
be invoked when we have already designated a function call as a
composable call.

Note that the previous annotator had extra logic to prevent recomputing
the same information mutliple times for a single file in each
highlighting pass. Since the new extension interface already passes us a
ResolvedCall object, we no longer have to worry about that.

Fixes: 189340127
Test: ComposableCallCheckerTest.testComposableCallHighlighting, ComposableHighlighterExtensionTest
Change-Id: Ie130c94139ba77122543e67726f00cab43c3fd90

GitOrigin-RevId: 7987d826d1ff88b91fb568cb5263ea9c0322a8c8