Update ChooseClassDialog to support the background PsiElement loading
In one of the latest platform updates PsiElementListCellRenderer has
been updated to load the UI elements in the background. This has caused
the initial measuring of the dialog to be wrong as it is using the
"loading" state and not the actual contents.
This change updates the ChooseClassDialog to update when the contents
have been loaded.

We only have to do this once since the main thing we care about is the
width of the cell. Once the first one is calculated, we can stop
listening.

Fixes: 248516787
Test: Manually verified
Change-Id: I8bdf3712994b0d09b3b5b21e5bdb5c2f57d6a038

GitOrigin-RevId: f901b8a1938395dad95a4dbf42d2071ca293c3d5