Handle exceptions in EmbeededCompilerClientImpl NonBlockingReadAction
Exceptions thrown during the execution of a NonBlockingReadAction will
be logged as errors and, after d747a2d, flagged to the user. In this case,
we want to handle the exceptions ourselves without the user necessarily
knowing about it since we have a retry logic and some special cases for certain
compiler errors.

Fixes: 253091994
Test: EmbeddedCompilerClientImplTest and manually verified
Change-Id: Ib529aea262201be8f5777d70c32b4f798188226b

GitOrigin-RevId: efbadab5abcc38b9cb1eb961ed3027ca07dbac2e