Handle Offline Device Better in ShowLogcat
Instead of showing an error dialog if a device is offline when a user clicks the link for a device that is offline, open a Logcat panel for the offline device. When the device connects again, the panel will attach to it. Like it already does when it restores state.

In order to do this, we need to pass in more information to ShowLogcatListener. This information is stored in ShowLogcatListener.DeviceInfo.

Bug: na
Test: Manually
Change-Id: I787f166ddfda7ffa0a178378ff1f45177f1b43a6

GitOrigin-RevId: a2dbf50f4957ba33159fc2d5b7e499bc5db829a8