[Dsl] Remove throwing of runtime errors where possible
Clients are unlikely to check the runtime errors, given that we don't
clearly document them.  Instead, either adapt internals to take a
recovery action in the circumstances that would have caused an
exception (e.g. on an invalid operation, leave the Dsl model
untouched); at some points on the interface, make return types
nullable.  (This is strictly speaking an incompatibility, but judging
by the Studio code that needed modification, precisely zero cases were
previously handling the exceptional situation, so in practice the
situation is not made any worse by this.)

Bug: 254475893
Test: existing tests adapted
Change-Id: Ifc9c8ca87f6164952a42bc8d7585401dbf7c45c3

GitOrigin-RevId: b1829b405e203afe4add2318aacf14559f09b4fa