Relayout the preview positions when reducing the window size
When extending the window size of design surface makes the viewport
of scroll pane become larger, which relayouts the positions of
previews.
However, when reducing the window size, the viewport size is not
reduced together -- it keeps the same size and showing the scrollbar
for changing the visible area. Since the size of viewport is not
changed, the layout manager doesn't relayout the previews.

In this change, add a listener to the container pane to make sure the
previews are relayouted when window size is changed.

Fixes: 222363553
Bug: 245363234
Test: manually (UI-related)
Change-Id: Id5bce398fb7283652767ec504cb84512b9649b17

GitOrigin-RevId: b2c49319d5d5d8cd6339022286aef3d67667e9f2