[Sync] Target JDK 1.8 for tooling-api Kotlin code
It doesn't seem to me that this is in fact enough in general: as far
as I can see, the Java files in this API module would still be
compiled with JDK 11 (class file version 55).  However, this change is
enough to allow Android Studio built and run from the IDE to
successfully import projects using AGP 3.5.0 / Gradle 5.5.  (No change
is needed for building from bazel.)

Bug: 251263800
Test: manual
Change-Id: If0e451322790daaa15c2dace9619560372cdd44e

GitOrigin-RevId: f7ee82ceee0b4b1f31faa5756d01e57faab5a9b6