Keep track of where AndroidSystem is initialized.
Sometimes in tests, AndroidSystem is initialized multiple times, which
is not allowed. This could happen inadvertently due to changes made
by others, so this change keeps track of where the AndroidSystem is
first initialized. When throwing the exception indicating this
exceptional condition, we can now include information about where it
was previously initialized.

Test: None
Bug: None
Change-Id: I08f34bf04a7a2f8b88bbaa9910598c518258a452

GitOrigin-RevId: d71418fd05d48bdc0ca34d627abea96b9bab1fb6