Remove unused functions in ForegroundProcessDetection
startListeningForEvents and stopListeningForEvents were originally
intented to start and stop foreground process detection based on the
state of the tool window: visible or not.

These two functions are not actually necessary, because Layout Inspector
is meant to keep working when the tool window is closed. A pop up is
shown when the tool window is collapsed, allowing the user to stop
Layout Inspector. This also stops foreground process detection.

Test: NA
Bug: NA
Change-Id: I7991baa1cb7fac49042dd6d388a75caafbc017d4

GitOrigin-RevId: 70fd4c2b2d1b5116fd16d9745c72dd1ed4b77a2c