Remove unused functions in ForegroundProcessDetection
startListeningForEvents and stopListeningForEvents were originally
intented to start and stop foreground process detection based on the
state of the tool window: visible or not.

These two functions are not actually necessary, because Layout Inspector
is meant to keep working when the tool window is closed. A pop up is
shown when the tool window is collapsed, allowing the user to stop
Layout Inspector. This also stops foreground process detection.

Test: NA
Bug: NA
Change-Id: I87ff518eaf11e23fd91f21fddca1edce5dc50a4d

GitOrigin-RevId: 61697074239873aab63bce00c4a1e214cd7cd8c4