[AUA] Upgrade project JDK if required by AGP
We only support running Gradle with a JDK version corresponding to the
version distributed with the Studio contemporaneous with the AGP
release.  (Other JDKs may work, but if they do they are unsupported.)

If a user upgrades over an AGP version at the support boundary,
attempt to update the project JDK to one that has the right language
level.  If none such is available on the user's system, block the
upgrade.

Bug: 251263800
Test: manual, unit tests added, existing integration tests
Change-Id: If8be38b1b6dcc6d5659622e9853bd559c296204a

GitOrigin-RevId: 93fb32e0e54be97932b3d8d3174b3cf4f7194569