[AUA] Simplify the AGP Upgrade Assistant toolbar entry
Although it is a refactoring action in some sense, it's not a
refactoring action in the sense of operating on its direct context,
which is apparently what BaseRefactoringAction is attempting to
simplify (to identify contexts in which the action should or should
not be offered).  Instead, simply inherit from AnAction.

Fixes: 256229057
Test: manual
Change-Id: Ie1ed484056d7971cf33079fffec770fc60c6d5c8

GitOrigin-RevId: f20c7bb43ac0b28af70a3f0607357468279d1f27