Mark Preview as out of date when resources are modified
This ensures the user is aware that resource modification requires a
full build.

This creates a new OutOfDate status that allow the Compose Preview to
tell the user the project is Out of Date. When changing resources,
Fast Preview can not help. We pass the information up to the Compose
Preview so the UI can signal to the user that the project is out of date
even if Fast Preview is enabled.

Test: ProjectBuildStatusManagerTest ProjectBuildStatusManagerResourcesTest
Bug: 237996794
Change-Id: Ic8d4ec3299d3dfa9c5550e4eb2dd736d9fb5c0d6

GitOrigin-RevId: 987bd3a8989704a6bbc1750785ba09eaaebcfaaa