Speculative fix for flaky test
This test is occasionally failing, due to an inability to resolve a
reference in the form `Manifest.permission.SOME_PERMISSION`. This
operation is done immediately after updating the manifest to include
that permission's definition.

One theory is that the update to the manifest has queued a write action,
which in most cases completes before the test verifies the reference but
occasionally does not. By surrounding the reference resolution in a read
action, I'm hoping that we can ensure any such action is complete before
we resolve the reference.

Unfortunately I'm not able to validate locally that this fixes the
flakiness, because I've not been able to repro the flakiness even with
a large number of runs_per_test. But this fix shouldn't make anything
worse, so seems worth a try.

Note also that I've updated the code to use myFixture.elementAtCaret,
which is more standard for tests. The comment saying that we can't use
it is out of date, because it now works perfectly for this case.

Bug: 256845633
Test: com.android.tools.idea.res.LightClassesTestBase
Change-Id: I2ecf8f1ab254aa2d14b91801695407ede6767eb9

GitOrigin-RevId: 3dd660410d8e6e935f92a817af9adc813b05c364