Do not automatically select tab 0 on shared tab removal
When removing the shared issues panel tab, do not force the selection of
tab 0 since this can change the tab the user had selected before.

This can not easily be verified from unit tests since the
MockContentManager has a different behaviour to the ContentManagerImpl
and will always select the first tab.
To test it, open two editors, a Kotlin file and a colors XML file.
Switching from the Kotlin file file to colors XML should not switch
tabs.

Test: Manually verified
Bug: N/A
Change-Id: I86ca1c70be78561dd33fddd19f52abe0409594fd

GitOrigin-RevId: b03d4950dbf11eb63bcd6b56abd0c1778a63d301