Create basic touch event test for Screen Sharing Agent
This change introduces a new test case for the Screen Sharing Agent
that tests how it responds to mouse clicks/touches. It also changes
the Event Logger app to print a log message when the app has resumed,
so that tests know if it is available to send touch input to yet.

Bug: None
Test: ScreenSharingAgentTest.kt
Change-Id: I49ba6a7ba17824e1009bf2e055a6003f269c1fca

GitOrigin-RevId: 4764a252bec8208a7ec2475510fd57caa9e3e997