Enable BuildConfig in some test projects
And update model snapshots now that BuildConfig is disabled by default

This change also modifies AndroidGradleTestUtils to allow tests to
specify whether to include BuildConfig source files (false by default).
This functionality is needed for the tests in
BuildConfigDefaultRefactoringProcessorTest.

This change also changes the version in
BuildConfigDefaultRefactoringProcessor.necessityInfo from 8.0.0-beta01
to 8.0.0-alpha09, because that will be the version of AGP in which
BuildConfig will be disabled by default.

Bug: 243130943
Test: existing
Change-Id: I06ea7220ece8fbdb9c6253001ee7a839dd1ee970

GitOrigin-RevId: b73d597cf888169826bac10505af79cb9c564445