Retry flaky operations in ScreenSharingAgentTest
Some operations are inherently flaky when interacting with the
emulator. Specifically, installing an app may fail if the device's
PackageManagerService is not ready yet, and touches may not be
received immediately after an app starts as it may not actually be
visible yet. Retry these operations a reasonable number of times in
order to decrease the overall flakiness of the test.

Test: ScreenSharingAgentTest.kt
Bug: None
Change-Id: Ib97dc29c797be61fe4ba81efb8a89e075f6c4043

GitOrigin-RevId: 3d8b80172d40153f0c9c81ba2c78e79bb5fdec65