Store global references to existing DeviceModel
This change introduces the storing of static references of references of `DeviceModel` because when multiple projects are open, they  need to coordinate with each other.

When multiple projects are open, they all share the same device, on which a thread is running to do foreground process detection.

When a project asks the device to stop foreground process detection, it stops not only for that project, but for all the others too.

On-device foreground process detection should be stopped only if the device is not the selected device on the `DeviceModel` for any open project.

Storing static references could be avoided by changing the communication protocol between Studio and device see b/257101182. We might consider doing this change later on.

Test: ForegroundProcessDetectionTest, DeviceModelTest
Bug: NA
Change-Id: Ic1eb91a21a8ac4bc759906c4486926e1891027c6

GitOrigin-RevId: 1088cc61aca3b3951e733236c43029a7b75bf63c