Move DescriptorRenderer code to new file
Part one of refactoring this code. I've moved/copied function
definitions to a new file, but have avoided modifying them. This has led
to a bit of duplicated code, but it will be removed in a subsequent
change.

I am not touching the corresponding tests in this change, so that they
continue to run as-is and ensure no regressions.

Bug: 257121497
Test: ComposeCompletionContributor
Change-Id: I329807687c45b9f26b9adf48467fe1b6cd3e6dd3

GitOrigin-RevId: 12ade57ad07cc5d8091df97d092a1b4b554a5070