Optimize package name query
Cache information on package names and AndroidFacets.
This code is on the hot path, invoked by the compiler
front end 1000s of times. Previously, every invocation
would allocate an array that could be 1000s of elements
for large projects, and facet iteration was inefficient.

This change adds a project-wide cache that store mapping
from package name to AndroidFacets with such name, and it
also stores scopes that should be used for the manifest
index when querying for the package name.

Bug: 248576926
Test: manually tested against 5500 Gradle subproject
      Android build
Change-Id: I21a5d9d46df6344ccc0c9492287a861dd20dcd8d

GitOrigin-RevId: 4492ce9bf834ddc9eb3c4bcdf7cefd8469687319