Add getOptions and addOptions to ProfilingConfigurations
This change is part of the transition to using technology-specific
option/config fields from the all-encompassing UserOptions
proto. The addition of the getOptions and addOptions on top of
the toProto method (converts to UserOptions proto) prepares
the codebase for the transition, allowing us to add and verify
the creation and setting of tech-specific options protos, while
preserving backwards compatibility the UserOptions provides us.

For clarification, getOptions is simply a helper method each
ProfilingConfiguration implements to construct an options
proto using the ProfilingConfiguration data. This is then used
in the addOptions method, which takes in a TraceConfiguration
and sets the respective options field.

Bug: 258515754
Test: added
Change-Id: I51f28720772857aab54a6681527053074a5d2108

GitOrigin-RevId: d321a05a6717584169803f3dce483613bd954a5f