[platform] IDEA-305295 Fix post-commit warnings
Make a field final, replace <Type> with <>.

Make getSelectedDirectories() return a not null array
instead of an array of not null elements. It could be
that both were meant to use, but it's unlikely, as it
delegates to methods returning non-null arrays
with no nullability specified for elements.

Remove the now unnecessary paneId field. That's
some unfinished business from my previous commit.

GitOrigin-RevId: 985a1c44d7d179e6165cf58947d2ef3c39bebf7f