Move updateVisibilityAndNotifications to an updates flow
This allows to conflate multiple requests and to handle the refresh in
one place.

The main reason to do this is so requests can be sent during the
deactivated state and handled later. Also this helps to solve the
problem where the refresh would happen before the workbench is
initialized correctly.

Test: Manually verified
Bug: 258284591
Change-Id: I165aead12afb17a653870497fb9ad74823664ac4

GitOrigin-RevId: dd10c8b0806b6dcdeb1a827ee60ebbe89e84b864