[AUA] Convert uses of GradleVersion for versions of Gradle
Rather than use the generic GradleVersion class from the Studio
common-sdk, attempting to represent general version string semantics,
use the specific org.gradle.util.GradleVersion class which represents
the semantics and syntax of versions of Gradle.

This change would be almost invisible were it not for the fact that
the org.gradle.util.GradleVersion toString() method does not simply
return the string form of its version, but prepends "Gradle "; for
programmatic use of the version string, we replace explicit or
implicit stringification with a call to getVersion().

Bug: 242691473
Test: existing, slightly strengthened
Change-Id: I8d3036ca9ea6d4f97a3c1258682be1111a906a28

GitOrigin-RevId: a67fed9f175f47e63cbffe7bb9a2cd03a6dbf51f