Ignore recomposition counts from framework composables
The recomposition counts are present to help users find problems
in their own code: not in the framework code.

There are framework code that generates unwanted recomposition
counts (see b/256028047). We do not want the image to highlight
these framework composables.

Add a flag such that recompositions in framework packages can
be turned on. This may be useful for the compose team.

Fixes: 257548903
Test: AppInspectionTreeLoaderTest
Change-Id: I9fdc56364db0219d7ff7bf1b57be64ee510b8097

GitOrigin-RevId: 733b988c3c7f511e6b2238018480ac91b3452855