Remove Dolphin_V1 from sync-perf-tests
... and add block-network to smoke tests so they behave closer to the
perf tests.

Dolphin_V1 started failing because it depends on ndk-21, but this
dependency was removed. We decided to no longer update the test to use a
different NDK version since there are other tests that cover sync V1 vs
sync V2 differences and Dolphin_V2 tracks performance of sync for native
projects.

Bug: 228208989
Test: This change removes a test and modifies others in sync-perf-tests.
Change-Id: I59f6ab91ed1d78d00254722db2188d57591b1fae

GitOrigin-RevId: 7da6479a3dd98d2da384e898e607e682e1ab065c