Add back slider to perfetto profiling config
In ag/20404069, PerfettoConfiguration was
removed as a child of AtraceConfiguration and thus
lost some inherited behavior. One of which was that
the buffer size would be determined via a slider in
the custom profiling configuration settings. This is
now added back so the user may customize the system
trace buffer size.

Fixes: 259285311
Test: manual
Change-Id: I1df7fea720fda0f4cefab4ac813637b9f78c7efe

GitOrigin-RevId: 64bc5efe2ed6489eb66381b9bb2fa34fc0ab6e1a