Fix post build model handling in wear run configurations
When `MakeBeforeRunTaskProvider` builds the target of a run
configuration it stores the result of the build in a user data attached
to the run configuration and it includes post-build models, which for
certain versions of the AGP are used to locate the output `.apk` files.

However, as it needs a run configuration which is a `UserDataHolder` it
historically casts the `RunConfiguration` to
`AndroidRunConfigurationBase` and this does not work with
`AndroidWearConfiguration`.

(1) Replace the cast to `AndroidRunConfigurationBase` with a cast to
    `UserDataHolder`

(2) Adjust `ProvidersTest.kt` to support `AndroidWatchFaceConfiguration`

(3) Add an `ApkProvider` test for `AndroidWatchFaceConfiguration`.

Bug: 248220889
Test: ApkProviderIntegrationTestCase
Change-Id: I10ddd3b49fa0b0963b09b5d49142d6529f6067a6

GitOrigin-RevId: 00c742571d1a8a00cc89cf4cc0a4a7d570603222