Ensure JfrReportManager isn't created early
By making JfrFreezeReports.freezeReportManager into a function, we
ensure that it's not create (and more importantly, that it doesn't
subscribe to performance messages) until it's explcitily requested.

Additionally moving the call to `createReportManagers()` to inside the
test for this feature's flag. As it currently stands, the report
managers are always being created and start recording data even when the
flag is off.

Bug: n/a
Test: Manually verified
Change-Id: I745da097aaf35a462882e8cac967bf818d4ae4bc

GitOrigin-RevId: e145b56fca7270c71691f525e6a5b5be12ab2289