Fix resource dependency set for androidTest modules
Module dependency sets for resources are supposed to include main
modules with Android resources on which the source module depends.

`androidTest` modules do not usually directly depend on other modules
but do it rather via a test dependency on the `main` source set.

Partially fix the problem by always starting dependency traversal from
the `main` module.

This does not fix the problem when modules are referenced via
`androidTestImplementation` which requires a proper dependency query be
used instead of simply starting the traversal from the `main` module.

Bug: 235374639
Test: n/a (to be added with the proper fix)
Change-Id: I35c286ff6533767e1b498bac554fb3aa7a6f2e1b

GitOrigin-RevId: 0b7e154b043f60dc632fbfe29b12decafd7b4314