Select preferred version if none found
When the IntelliJ platform requests adding an unknown library dependency
to an Android module Android Studio ignores the requested library
version even when it cannot find the right version to add. This
behavior results in a broken build configuration.

When adding a library dependency to Android do not ignore the supplied
preferred version if there are no better alternatives.

Bug: 227021560
Test: n/a
Change-Id: I45a4321f4b5c33e977d6be281461aace6170ad8a

GitOrigin-RevId: 75c620f379d85d13cb12b935bb00d9d69ab870e7