Add JfrTypingLatencyReports
This change adds a report generator using the JFR infrastructure to
gather callstack data when high typing latency is encountered.

I've deferred creating server flags until after this is reviewed, and
will add a subsequent change to do so. This is just in case the review
uncovers the need for more and/or different flags, to avoid churn. As
the code is written now, the new reports won't be enabled until I've
added a flag and start creating the appropriate report manager.

Bug: 259447928
Test: JfrTypingLatencyReportsTest
Change-Id: If90d3f95267c4f0a22bbbf984d1bfce034449646

GitOrigin-RevId: 28c69840985bbb8e296af4bab8559bc9756c3110