Created a dedicated studio ComboBox for Gradle JDK dropdown settings
This was necessary to have a dedicated studio component,
allowing us to have our customizations. At the same time be able
to extend the functionality like displaying the JDK path in the
dropdown but also support custom gradleJvm macros like #JAVA_HOME.

With the previous implementation, a JDK table entry was created for
all SdkItem specified on SdkComboBoxModel. We want to avoid this
behaviour in case of macros like #JAVA_HOME, for this reason now
it's possible to add SdkReferenceItem that are just simple references
without being necessary to create a dedicated JDK table entry
i.e: jbr-17 Jetbrains Runtime version 17.0.4 /user/jdks/jdk-17

As a note, the platform doesn't allow an easy extension of this part
having most of the constructor's package local.

Bug: 255545053
Test: N/A
Change-Id: I0fa036ec2a3a676fc73767eac68016aad794b4f5

GitOrigin-RevId: ff8c44b2fd80ef77d054f7d77b80c6d85cbc0dee