Fix border paint in Compose screen view provider
In change Iaf488f, we changed the shape policy to Compose screen view
providers. This inadvertently broke the Compose preview borders when
`showBackground` is set to true, because in this case BorderLayer will
paint `myScreenView.getScreenShape()` instead of the actual border. This
behavior was introduced in change I38811, but the intention was to only
affect vector drawables. This change adds an extra field to BorderLayer
to make it possible to force painting the border for Compose Previews.

Fixes: 245826837
Test: N/A. Visually verified the border is painted correctly.
Change-Id: Id4e954982153dd26eac9e39af42d143c064f5201

GitOrigin-RevId: 14ceb78cd9ef64f8dc470de130371640796edd12