Do not report cancelled Visual Lint tasks to crash reporter
Currently, when visual linting tasks are cancelled because new ones have
been scheduled, they trigger a crash reporting. That should not be
happening as this is not a crash but a normal behaviour of the feature.
The same is true for render tasks evicted from the queue because it is
full.

Bug: 256698778
Test: checked manually
Change-Id: I6752826b298012bb2560ba709bf08fdbf675c36b

GitOrigin-RevId: 2ba3b21c53627d5a575cc5bec6b06bc32ce7669f