Replace fromTypeAndMode with fromTraceConfiguration
Because the new unified trace configuration does not
have a top-level field for trace mode (as this field
only matters in art options). The fromTypeAndMode was
changed to be based on the overarching TraceConfiguration
so that it could determine which option is set. And, in
the case of art options, will determine and return the
correct ProflingTechnology based off its TraceMode.

Bug: 258515754
Test: added
Change-Id: I00fdb5e683b259b4be11aa2b164dd73b5188d6d0

GitOrigin-RevId: f6bb73237f94778ed25249cbef4073a8fdb4335b