[BA] Fix ignored flaky DownloadsAnalyzerTest
Try to remove changing gradle home path that seem to be the reason of
the flakes. This is only needed for local runs. If this helps we will
need to find another workaround.

Also fix leaked threads error appearing in logs. Apparently the way
servers were registered as disposables was not correct and they were
closed after the check for leaked threads.

Bug: 240887542
Test: This fixes test
Change-Id: Ib394e62395dc2baf7ddd5ccb19f67b936b2c1636

GitOrigin-RevId: fa4555c2af6c1a9261d11903596c395b2092f705