Ensure GradleModuleModel is compatible with...
both pre and post eef8888f3c70d4cd4b3cbc306b470093d7bff353
serialization formats. We add back a removed fields myGradlePlugins and
use its contents if not null to figure out values for mySafeArgsJava and
mySafeArgsKotlin.

We mark getGradlePlugins() as deprecated but leave it in for downgrading
from models where it is not populated.

Bug: 260621829
Test: existing for functionality none for update issue
Change-Id: I6632b49b4cbc49509db1d0ddf7291376dd27ea5d

GitOrigin-RevId: 086cf86e504e0047d9e6630409791c15934745eb