Remove toProto and buildUserOptions methods
In a larger effort to remove UserOptions and its
usages, the toProto (and its helper method buildUserOptions)
have been removed as well as their usages. The only non-unit
test usages were found in the ProfilingConfiguration toProto
and equals methods. However, with the new unified trace config,
we were able to use the technology-specific options proto instead
of the UserOptions proto naturally.

Bug: 258515754
Test: updated
Change-Id: I30cfe820ea6be433c5085134c964a5c3cce1bc03

GitOrigin-RevId: 0dd17e7da970024ae33a33f0f0a147c8b6fa2b6c