Remove UserOptions from profilers cpu tests
As part of the larger effort to remove UserOptions,
the instances of UserOptions have been removed from
the top level profiler cpu tests. This change is not
harmful as, if needed, the unified trace config option
protos have been already been added in prior commits
to preserve functionality and test passage.

Bug: 258515754
Test: updated
Change-Id: If34018de61855d468c7a781dd498f23135f9592f

GitOrigin-RevId: 15d21c5050a5ef68b66e8f27888095da0d349d52