[IconPickerDialog] Improve metadata file handling
We now use a more consistent and safe pattern to read and write a
MaterialIconsMetadata file. Which addresses a missing `close()` call and
using a large String to write into the file.

The logic for
- copying metadata for the first time
- updating metadata file

Have been updated to reduce write operations while being cancellable.
These processes should now be significantly faster.

Modified backdroundExecutor disposing logic to wait a couple of seconds
for termination, so that all resources are likely to be released within
the Dispose call.

Bug: 253877925
Test: updated
Change-Id: Ia1c197aa1ec35c43b5cb56a27600d64c78ecfb34

GitOrigin-RevId: 7848194a0cbdfdefc4377c4162797ebc19a6c3f3