Implement listener to track ProjectView selections
On register of project tool window subscribe listener to its content
selection changes. We then track fist removed from selection page ID
and last received selection event time.
On remove selection we check when was the last event, if it happened
long enough ago to treat this as a new session - we replace first id.
On add selection event we schedule a runnable to run after 10s and
this runnable keeps the time it was added. When it is executed it
first checks for last known event time, if it is different from its
own time this means something else happened already and it should not
proceed.
This cool down is required to eliminate from reporting accidental
switches as well as automatic switches during panel init process.

Listener initialized in a special 'init' ad schedules state report
right away, this allows us to track initial previously saved setups.

If 'from' field is Android view then trigger a survey on this.
Survey is controlled via feature flags and there are internal
mechanisms to not annoy the user, it is only shown once in a while.

Bug: 258731735
Test: Added ProjectViewListenerTest
Change-Id: I9a1a7e03ba9fc15e93ab0987ce7c16fe64ddf547

GitOrigin-RevId: c7ef609910934ba56568b1c3667e509bcf21ce13