IDEA-307652: Disable Device Mirroring feature until libraryscreen-sharing-agent.so lib is provided
Device Mirroring in Android projects is disabled due to missing libscreen-sharing-agent.so library.

The feature is disabled by changing `DeviceMirroringSettings` `deviceMirroringEnabled` and `synchronizeClipboard` flags values to `false`.
In addition to that, the UI settings option that can modify the Device Mirroring feature state is hidden.

The Device Mirroring feature will be re-enabled after the Android Studio EE Beta 4 sources are merged and missing library is provided.

GitOrigin-RevId: ca0995334ff027a66cdf1eb893f22df1781cddc2