Remove system trace slider & set const buffer
It was discovered that while buffer size is configurable
via a slider in studio-side, this value is not actually
used as a hard-coded buffer value is used on base side.
However, it was decided that the user should not be able
to configure this value. Thus, we can remove any code that
makes it seem like the user can. This includes the slider.
Along with the aforementioned decision was that the default
buffer size should thus be 4 MB for system trace.

Bug: none
Test: none
Change-Id: Ibe9b117559548b7fba73eda3dc877d9681377a32

GitOrigin-RevId: e97b85c9569b4e062e6e23f5eb25ae53a69a32aa